requirements.txt
openai-whisper
coqui-tts
openai
sounddevice
numpy
python-dotenv
